Client and server can both send information
As it stands right now, the client and server can both communicate on an
open socket for as long as the socket is kept open. However, once the
client closes the socket. The server must be restart to accept any new
connections. The reason for this at this point, it that the server has
no way of knowing the socket has been closed (as far as I can tell).
Unless the client actually sends a graceful shutdown message, we won't
know until we actually go to write to the socket.